backend/requirements.txt
Flask
requests
flask-cors
python-dotenv
uvicorn
